Proposal: extensible metadata API
This change proposes an extensible way for a ledger to expose metadata
about the token to facilitate integration with wallets and user
interfaces: name, symbol, the number of decimal places for user display,
icon, official web-page, etc.

The new API makes some of the previous methods (`symbol`, `name`,
`decimals`) excessive, so we remove these methods from the standard and
replace them with standard metadata entries.