Proposal: Rename principal in `Account` type (#32)
* Update .did file with changes from #26

With #26 we created the Account and Subaccount types in the README.
However, the .did file wasn't changed. This bring it into alignment.

* Rename "of" to "principal"

The `Account` type is a combination of a principal and a subaccount. The
subaccount was correctly called "subaccount" but the principal was
called "of" for some reason. In the context of the `icrc1_balance_of`
method, this makes sense. However, in the context of `icrc2_transfer` it
does not.

In the future we may implement even more methods where "of" doesn't make
sense as the name of the principal. By separating the usage from the
definition, `Account` will remain flexible for all these use cases.

* Add quotes to principal type

"principal" is a reserved keyword. Wrapping this field in quotes
indicates it is a label and not a type.